nuevos cambios de vista

+++ Models/DetalleCompra.cs
@@ -6,12 +6,14 @@ using System.Threading.Tasks;
 
 namespace Trabajo_Final.Models
 {
+    
     [Table("t_detalleCompra")]
+    
     public class DetalleCompra
     {
-        public long Id { get; set; }
-        public long CompraId { get; set; }
-        public long ProductoId { get; set; }
+        public int Id { get; set; }
+        public int CompraId { get; set; }
+        public int ProductoId { get; set; }
         public Producto? Producto { get; set; } 
 
         public int Cantidad { get; set; }
